use new _at_fork_reinit() for lock
introduced in python 3.9: https://bugs.python.org/issue40089
and requiered (e.g. by the logging library)

This was also fixed in
 #90 / 46d71bd0b1f869e0c60992225334a0bcddbaac7c
while this commit was still in a WIP branch, and merged before I could intervene ;-)
This removes checking for an exception, again, because the member's function should be present when the outer function is called (i.e. in python >= 3.9)